requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
chromadb==0.5.5
sentence-transformers==3.0.1
transformers==4.44.2
torch>=2.2.0
numpy==1.26.4
scikit-learn==1.5.1
python-docx==1.1.2
pymupdf==1.24.9
nltk==3.9.1
